Refactor: Move DSA signature related methods to DsaSignatureValidation
